[CORE] Extract Info Exchange out of VersionManager
Created InfoManager to divide responsibilities of VersionManager.
