+++ core/META-INF/MANIFEST.MF
@@ -29,6 +29,7 @@ Export-Package: com.thoughtworks.xstream,
  saros.account,
  saros.activities,
  saros.annotations,
+ saros.communication,
  saros.communication.chat,
  saros.communication.chat.muc,
  saros.communication.chat.single,
