+++ server/src/saros/server/net/ServerFeatureAdvertiser.java
@@ -1,8 +1,8 @@
 package saros.server.net;
 
 import saros.annotations.Component;
+import saros.communication.InfoManager;
 import saros.preferences.PreferenceConstants;
-import saros.versioning.VersionManager;
 
 /** Instructs the network layer to advertise this Saros instance as a server. */
 @Component(module = "server")
@@ -11,9 +11,9 @@ public class ServerFeatureAdvertiser {
   /**
    * Initializes the ServerFeatureAdvertiser.
    *
-   * @param versionManager to add info
+   * @param infoManager info manager to add info
    */
-  public ServerFeatureAdvertiser(VersionManager versionManager) {
-    versionManager.setLocalInfo(PreferenceConstants.SERVER_SUPPORT, Boolean.TRUE.toString());
+  public ServerFeatureAdvertiser(InfoManager infoManager) {
+    infoManager.setLocalInfo(PreferenceConstants.SERVER_SUPPORT, Boolean.TRUE.toString());
   }
 }
